+++ crates/oxc_macros/src/declare_oxc_lint.rs
@@ -1,43 +1,16 @@
 use convert_case::{Case, Casing};
 use proc_macro::TokenStream;
-use quote::{quote, ToTokens};
+use quote::quote;
 use syn::{
     parse::{Parse, ParseStream},
-    token::{SelfType, Token},
     Attribute, Error, Expr, Ident, Lit, LitStr, Meta, Result, Token,
 };
 
-enum IdentOrSelf {
-    Ident(Ident),
-    SelfToken(SelfType),
-}
-impl Parse for IdentOrSelf {
-    fn parse(input: ParseStream<'_>) -> Result<Self> {
-        if input.peek(Token!(Self)) {
-            // input.parse::<Token!(Self)>()?;
-            Ok(Self::SelfToken(input.parse()?))
-        } else {
-            Ok(Self::Ident(input.parse()?))
-        }
-    }
-}
-impl ToTokens for IdentOrSelf {
-    fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
-        match self {
-            Self::Ident(ident) => {
-                ident.to_tokens(tokens);
-            }
-            Self::SelfToken(s) => {
-                s.to_tokens(tokens);
-            }
-        }
-    }
-}
 pub struct LintRuleMeta {
     name: Ident,
     category: Ident,
     /// Struct implementing [`JsonSchema`] that describes the rule's config.
-    schema: Option<IdentOrSelf>,
+    schema: Option<Ident>,
     documentation: String,
     pub used_in_test: bool,
 }
@@ -63,14 +36,7 @@ impl Parse for LintRuleMeta {
 
         let schema = if input.peek(Token!(,)) {
             input.parse::<Token!(,)>()?;
-            // input.peek(Ident).then(|| input.parse()).or_else(||
-            //     input.peek(Token!(Self)).then(|| input.parse())
-            // ).transpose()?
-            if input.peek(Ident) || input.peek(Token!(Self)) {
-                Some(input.parse()?)
-            } else {
-                None
-            }
+            input.peek(Ident).then(|| input.parse()).transpose()?
         } else {
             None
         };
@@ -100,18 +66,21 @@ pub fn declare_oxc_lint(metadata: LintRuleMeta) -> TokenStream {
     let import_statement = if used_in_test {
         None
     } else {
-        Some(quote! { use crate::rule::{RuleCategory, RuleMeta, RuleSchema}; })
+        Some(quote! { use crate::rule::{RuleCategory, RuleMeta}; })
     };
 
-    // let schema =
-    //     schema.map_or_else(|| quote! { None }, |s| quote! {
-    //     Some(gen.subschema_for::<#s>()) });
-    let schema_impl = if schema.is_none() {
-        Some(quote! {
-            impl RuleSchema for #name {}
-        })
+    let schema_impl = if let Some(schema) = schema {
+        quote! {
+            gen.subschema_for::<#schema>()
+        }
     } else {
-        None
+        quote! {
+            {
+                let mut obj = SchemaObject::default();
+                obj.object().additional_properties = Some(Box::new(Schema::Bool(true)));
+                obj.into()
+            };
+        }
     };
 
     let output = quote! {
@@ -127,8 +96,6 @@ pub fn declare_oxc_lint(metadata: LintRuleMeta) -> TokenStream {
             }
         }
 
-        #schema_impl
-
         impl schemars::JsonSchema for #name {
             #[inline]
             fn schema_name() -> String {
@@ -143,21 +110,12 @@ pub fn declare_oxc_lint(metadata: LintRuleMeta) -> TokenStream {
             fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
                 use schemars::schema::{Schema, SchemaObject};
 
-                let mut schema: Schema = #name::schema(gen)
-                    .unwrap_or_else(|| {
-                        let mut obj = SchemaObject::default();
-                        obj.object().additional_properties = Some(Box::new(Schema::Bool(true)));
-                        obj.into()
-                    });
+                let mut schema: Schema = #schema_impl;
 
                 let schema = match schema {
                     Schema::Object(mut obj) => {
                         let meta = obj.metadata();
                         meta.title = Some("Config for ".to_string() + Self::NAME);
-                        // meta.description = Self::documentation().map(Into::into);
-                        // if let Some(docs) = Self::documentation() {
-                        //     obj.extensions.insert("markdownDescription".into(), docs.into());
-                        // }
                         Schema::Object(obj)
                     },
                     s => s
@@ -187,7 +145,3 @@ fn parse_attr<'a, const LEN: usize>(
     }
     None
 }
-
-// fn parse_comma() {
-//     input.parse::<Token!(,)>()?;
-// }
